+++ packages/amino/src/secp256k1wallet.ts
@@ -1,5 +1,5 @@
 import { Secp256k1, Sha256 } from "@cosmjs/crypto";
-import { Bech32 } from "@cosmjs/encoding";
+import { toBech32 } from "@cosmjs/encoding";
 
 import { rawSecp256k1PubkeyToRawAddress } from "./addresses";
 import { encodeSecp256k1Signature } from "./signature";
@@ -34,7 +34,7 @@ export class Secp256k1Wallet implements OfflineAminoSigner {
   }
 
   private get address(): string {
-    return Bech32.encode(this.prefix, rawSecp256k1PubkeyToRawAddress(this.pubkey));
+    return toBech32(this.prefix, rawSecp256k1PubkeyToRawAddress(this.pubkey));
   }
 
   public async getAccounts(): Promise<readonly AccountData[]> {