+++ scripts/wasmd/send_first.js
@@ -3,7 +3,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 const { coins } = require("@cosmjs/amino");
 const { Random } = require("@cosmjs/crypto");
-const { Bech32 } = require("@cosmjs/encoding");
+const { toBech32 } = require("@cosmjs/encoding");
 const { DirectSecp256k1HdWallet } = require("@cosmjs/proto-signing");
 const {
   assertIsDeliverTxSuccess,
@@ -23,7 +23,7 @@ const faucet = {
 async function main() {
   const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic, { prefix: prefix });
   const client = await SigningStargateClient.connectWithSigner(rpcUrl, wallet, { prefix: prefix });
-  const recipient = Bech32.encode(prefix, Random.getBytes(20));
+  const recipient = toBech32(prefix, Random.getBytes(20));
   const amount = coins(226644, "ucosm");
   const fee = calculateFee(100_000, "0.025ucosm");
   const memo = "Ensure chain has my pubkey";